Renamed BPM to bpm and fixed the get_song_id test
