+++ .gitignore
@@ -1,7 +1,9 @@
+### Excludes ###
 music/*
 .vscode/
 .idea/
 venv/
 *.pyc
 
+### Includes ###
 !*/.gitkeep
